TGBOT-238 move class to util directory
